Refactor render-pdf into render-core
This huge refactor changes the previous render-pdf into a format
agnostic rendering system for the reader.

Main changes include the renaming of PdfState -> DocState, load_pdf ->
load_doc and so on.

From now on render-core would handle the core rendering logic of the
reader in a format agnostic manner, whie other modules in the future
such as render-pdf.c or render-epub.c would handle format-specific
operations such as annotations and so on.